Moved files around and created main view instead of having it in the app.js
